backend/requirements/dev.txt
-r base.txt

black==19.10b0
flake8==3.7.9
flake8-black==0.1.1
pytest==5.3.5
pytest-django==3.8.0
pytest-cov==2.8.1
